Reload polldevs.cf on set intervals
This restructures the logic that was previously countained in zino.main.
Loading/reloading of polldevs.cf is now a scheduled, re-occurring job,
which will also take care of adding new devices and de-scheduling
deleted ones.

It also adds the concept of the state module, which is meant to
eventually hold global state for the Zino process.

It does not yet detect devices whose polldevs config has changed
enough that it warrants rescheduling of it.