Move DEFAULT_INTERVAL_MINUTES to config models
The default value was already in the PollDevice model. This ensures
the model and everything else gets the default value from the same
place.

This also changes the default to 15. While Zino 1.0 defaults to 15, it
has been running with a value of 5 minutes in production for years and
years.

We may need a mechanism to rather use whatever was given as the default
value in the config file, though.