Fix bugs uncovered by tests
Without default values, instantiation of Event objects would require
values for the log and history attributes.

Although it seems counter-intuitive and plain wrong for a long-time
Python developer, using empty lists as default values for Pydantic
models is OK, as Pydantic contains the magic to make copies of the
defaults on every instance.