Make Events registry into a Pydantic model
The registry it self needs to be serializable.  Attributes that are to
be serialized should be public, while purely internal attributes should
be marked as private.

The actual list of events and the highest running event id needs to be
serialized, but the internal lookup index for events by key should not
be serialized (or it will basically just be a copy of the events list,
which makes it 2x the size).