Restructure event types to subclasses of Event
Not all event attributes are common to all event types.  This ensures
a cleaner separation of concerns when it comes to the data model.

This obsoletes the EventType Enum, replacing it with subclasses of
Event.  Additionally, the `event_type` attribute was renamed to `type`
to be in line with the existing Zino 1 line protocol.  While `type`
alone would shadow the builtin `type`, as a class attribute it should
be fine (and we won't have to do fancy mapping when we get to
implementing the line protocol)