Rebuild Events index when constructing new object
When an Events object is constructed from a file dump (by way of the
Pydantic method `Events.parse_file`), it would start with an empty
index (since the index was updated every time its own methods were
used to create new events).  When the index is empty,
`get_or_create_event()` ` would not be able to look up existing matching
events, and we would fill the event registry with duplicate events.

This ensures the index is rebuilt from scratch when a new Events object
is instantiated.