Add sparsewalk implementation
Based on the new getbulk implementation, this sparswalk implementation
can return multiple table columns in a single operation.  The result
value is a dictionary keyed by the row index (as an OID value), where
each value is a dictionary with the keys being the column names of each
of the queried columns of the table.