Replace IntEnum with StrEnum for interface states
We do not actually ever need the numerical values from the MIB in our
state handling code, as the values are translated to strings for us by
the MIB layer.  It is therefore better to use a StrEnum, which makes
translation from MIB string values easier, and makes the saved state
even more human-readable.

Also, at least one of the states that Zino 1 operates with, "adminDown"
isn't actually a state value from the MIB.  It is used to flag the case
where an interface operational state is down not because of an error or
missing link, but because the administrator has disabled the port on
purpose.

Because of this, the InterfaceOperState enum was also renamed to
InterfaceState (since it doesn't necessarily represent just operstate
values).