Fix error_indication typehint
As demonstrated in the code, error_indication can
apparantly be a errind.RequestTimedOut object, which is a subclass of
errind.ErrorIndication. So im guessing any type of errind.ErrorIndication
could be used there.

Im leaving string as union, because the docstring for getCmd does state
it is supposed to be a str value, so im leaving it open for now