Fix broken exception handling in command dispatch
The command dispatcher stopped working as intended as soon as the
protocol class switched to using async methods as command responders.

This fixes the problem, while adding some tests for the intended
behavior: If a command responder raises an unhandled error, the
protocol class should log the exception details, respond with a
non-detailed error to the external client, and move on to serve the
next command.