Copy signature of wrapped function
For multi-argument functions, the command dispatcher needs access
to the wrapped function's signature to know the real number of
desired arguments,.

The downside to this is that the required `case_id` argument is listed
to a connected client as `event` instead.  Although it doesn't matter
much for the protocol itself, it could confuse a human connecting
to the Zino server.