Observe event changes and reschedule state dumps
This adds an event change observer to the `Events` state model.  This
observer ensures the state dump job is rescheduled if there's more
than 10 seconds until its next scheduled run.  This is to ensure we
commit state to disk "often enough" when things change, but not too
often.

This also sets the default interval of the state dump job to that of
the default polling interval.  Although Zino1 actually uses the value
from polldevs.cf for this, polldevs values aren't actually easily
available to the loop setup function, since the first polldevs read
happens in another job.