Change event creation/checkout/commit workflow
This changes the workflow for creating new events.  New events will not
be committed to the event index immediately, and will have an empty
id value.  The caller must explicitly use the `commit()` method to
ensure a new event is assigned an id and committed to the index, just
as it must use `commit()` to ensure any changes to existing events are
committed back to the index.