Update all callers to Events.get_or_create_event
The return value no longer includes a `created` boolean.  When
`event.id is None` it should be assumed the event was created/is new.

Also, it is not the responsibility of these callers to explicitly set
the event state to `OPEN`.  This is accomplished automatically when
a new event is committed for the first time.