Update all tests to use new commit workflow
This updates all tests that were failing because they were not using
the new checkout/commit paradigm for the event index.

All tests must assume that anything returned from `Events.get_or_create`
must be explicitly committed back to the index if they are new or
changed.  Also, tests cannot assume that an object return by `get` or
`get_or_create` is identical to an object the test is holding a
reference to, since the `get/get_or_create` will create checkout
copies.