Add address ignore mechanism
The original Zino Tcl code includes lists of addresses and address
patterns that should be ignored from router address collection.
This reimplements that.

The addresses should really be part of a config file (that we don't
have yet), but since the original Zino hardcodes it,
so will we for the first Zino 2 release.