Change PortOrIPAddress to SubIndex
In reality, this type is used for identification of sub-components of
routers.  It's a type that can be used for sub-indexing of events, in
order to be able to distinguish multiple events of the same type of on
the same router.

This gives the type a more generic name that explains what it's for.  It
also adds `None` into the `Union`, so that `Optional` can be dropped
from all places that use it: `None` *is* a valid `SubIndex` value for
reachability events, for example.