Stop erroneous overloading of port values in Event
Not all events should have a `port` attribute.  In fact, only the
`PortStateEvent` should have one, according to the Zino 1 API.

However, all event types, except for `ReachabilityEvent`, should have an
attribute that uniquely identifies a sub-component of the router the
event pertains to.  This is the value we need to make a unique
`EventIndex` value for events.  However, since this attribute is named
differently on each event type, we cannot rely on the `port` attribute
to give us this value.

This adds a `subindex` property to the `Event` class, which subtypes can
override to return the value of whichever attribute the subtype
considers its identifying attribute.

Not all existing event types had such an identifying attribute, as they
relied on erroneously using the inherited `port` attribute for it.