Fix incorrect descr value of PortState events
This fixes a small bug that set the `descr` attribute of PortState
events to a port's `ifdescr` value.  The correct value, according to
Zino 1 code, should be the port's `ifalias` value, which is sometimes
called a `descr` or "port description": It's the mutable port
description set by a router admin.