Replace `port` with `subindex` as index name
This removes the last vestiges of `port` as the name of an event index
component.  The `port` term was heavily overloaded to mean several
different things in the Zino 1 codebase, and we initially ran with that
idea.  This replaces the term with the more generic `subindex`, which as
already typed as `SubIndex`.