Protect against unset server variable
If this `Zino1ServerProtocol` instance was created with an empty
`server` argument, `self.server` will be `None`.  The protocol should
still work, but no notification channels can be discovered.  This will
respond to the client with the natural consequence of this, rather than
crash with an "internal error".

As noted in code review.