Clean up tests and types for peer_name
As mentioned in code review, the tests for `peer_name` for both the
server protocol and the notification protocol could both be named
better and have more explicit assertions for `None`.

Additionally, the `peer_name` properties of the protocols had incorrect
return type annotations, which did not allow for `None` values, so
this is also fixed.