Ignore trailing garbage args on API commands
Apparently, legacy clients may attempt to send extra garbage arguments
at the end of commands like the `USER` command, and Zino 1 was fine
with this, while Zino 2 would respond with an error, causing the
client to fail on login.

This changes the behavior of the legacy protocol implementation to
allow, but ignoree and log extraneous command arguments.