Ensure return value is awaitable on sad path
All API command responders are async functions.  However, this decorator
is not async and would, as a result of its exception handler, return
a non-awaitable result that caused strange behavior in Zino.

This ensures that whenever an API client inputs a non-existent caseid
to an API command, the return value (which is no longer the return
value of the decorated async function) should still be awaitable.  This
is ensured by explicitly returning a Future that has already has a
result.