Test for single-packet multi-line input
Apparently, multi-line input sent as a single packet will confuse
the `data_received` function of the protocol.  This tests for this using
the multiline input function.  However, the test will "hang" because
multiline input will still be waiting for the single line containing a
dot, so pytest-timeout was added to ensure the test times out if it
doesn't succeed very quickly.