Update test suite to use `message_received()`.
Now that we have the higher-level of abstraction of
`message_received()`, this is more readily employed by the parts of the
test suite that is only concerned with high-level testing of the
protocol implementation.

Also, `Protocol.data_received()` is not defined as an async function,
which our implementation sometimes could be used as.  The reimplemented
`data_received()` will never return a Future, so it cannot be awaited,
causing most of these high-level tests to fail.