Word default/override possibilities more explicit
Some feedback we got was that the example config file was missing an
example that showed that default values could be overridden in
individual router entries.  That doesn't seem to be the case, there is
an explicit example of setting a different community in the last router
entry.  This just changes the wording of the comments somewhat in a
feeble attempt to make things clearer.