Revert codecov workflow changes
These changes did indeed run the codecov uploader in a different
workflow, but CodeCov seems to have issues with connecting the uploaded
coverage data with the correct PR.  Not sure why, but we should rather
experiment with this in a throw-away repo / fork until we get it right.

Reverting the last four commits until we've figured things out.