Change TrapObserver protocol into full class
We want as little boilerplate as possible to create a new trap observer.
An easy way to do this is to make them proper classes, enabling them to
declare which traps they want to observe, and adding auto-subscription
functionality for loaded subclasses.

This changes the TrapObserver to be an inheritable class that takes the
running ZinoState object (because many trap observers will expect to
inspect and potentially update Zino's running state) and event loop as
constructor arguments, and changes the `__call__` method into the more
explicit `handle_trap` method, which now only requires the trap message
itself (since more long-lived data is now stored as instance
attributes).

This also updates corresponding tests