Add reason attribute to PortStateEvent
Some link state traps can communicate a reason (a string) explaining why
an interface changed states.  This is communicated through the `reason`
attribute of portstate events in legacy Zino, so we should add that also
in Zino 2 (even though we do not currently read any "reasons" from a
trap, since the supproted `locIfReason` variable is potentially obsolete
in 2024).