Add polldevs config to trap observers
The observer for link traps needs access to configuration information
from zino.state.polldevs.  In order to avoid having trap observer
instances access global state directly, this enables them to have a
reference to that state injected as they are initialized (i.e. the same
pattern as with the persistent state structure).