Make schedule_verification_of_single_port public
This makes `LinkStateTask._schedule_verification_of_single_port` a
public method so that other parts of the code can easily schedule
single interface polls.

It also expands the method signature for more usefulness: Deadline is
added to control how far into the future the job should run.  Reason
is added to be able to modify how the job is named in the scheduler
(so it can be more easily distinguished from other verify jobs)