Add initial models/algos for flap tracking
This is an initial reinterpretation of what Zino 1 is doing to track
flapping states, and is still subject to change.  This suggests
keeping a separate model for flapping state tracking, so that algorithm
methods can be concentrated in that model.  Flapping states could
also potentially be tracked directly on Port model objects, but we'll
try this version first.