Return early when no flapping state is known
If there is no flapping state tracked for an interface, it is definitely
not considered to be flapping yet.  Let other parts of the code
initialize flap tracking data for an interface once traps start
coming in.

This came after discussions with Håvard E about the original Tcl code,
which works on a confusing set of global variables.