Add Responder NamedTuple
In order to change the dispatch mechanisms of the API to support
subcommands, we want to switch to matching commands using patterns
instead.

This rewrites `_get_all_responder()` to build a dictionary of command
names to Responder objects, and also to build the corresponding regexp
match pattern for each command.