Rework `_get_responder` implementation
This reworks `_get_responder` to use pattern matching to match the
incoming API message to one of the known responders in the cached
responder dictionary.

For sub-commands, multiple patterns may match, so this takes care to
always select the longest match first, to ensure the sub-command and not
the "parent" command is matched.

When a command has been matched, the remaining part of the message is
split into arguments, which are returned as a list.