Remove ifDescr/ldescr comments.
This was resolved with Håvard.  ifDescr or ifAlias is not normally sent
with a trap, and Håvard can think of no good reason for needing to keep
a trap-specific mapping of ifindexes to ifdescrs or ifaliases.

Which means we can keep using our existing mappings and ignore these
comments.