Fix broken flap stats serialization
Tuples as dict keys cannot be serialized to JSON by Pydantic without
custom logic.  I.e. it serializes to a string representation, with
tuple elements separated by commas, but Pydantic is unable to
deserialize this representation as a tuple, causing Zino to crash
when loading stored state.

This annotates the PortIndex type with modifiers that let's Pydantic
successfully serialize and deserialize PortIndex dict keys to/from JSON