Clean up broken flap state updates
Flapping state stats were constantly being reset every time an event
was updated, due to a bug in the implementation.

This renames `FlappingState.flapping` in an attempt to clarify that
this is a flag used to indicate that a flapping state has been declared
(i.e. an event has likely been created), while the `is_flapping()`
method is only used to verify that the current stats qualify as flapping
and that a flapping state should/could be established.