Factor out event handling from link transition
Now that the logic is better understood, this factors out the link
transition code that updates events for interfaces in either a
flapping state or a non-flapping state.

There were some redundancies that were copied directly from Zino 1,
these have been factored out.  Also, the code was made slightly more
readable by fetching the associated `FlapState` objects if they exist.

Additionally, this fixes a discovered problem in the first version of
the code, in where a transition out of a flapping state wasn't handled
properly, because the port was looked up incorrectly in the index of
flapping states.