Make `BGPPeerSession` fields optional
All the values associated with a BGP peering session aren't necessarily
always available to us.

Case in point:  If a BGP transition trap is received, it will only
contain a peer IP address and the new operational state value of the
peering session.  We might not yet have learned the full details of the
peering session through SNMP, but now we want to register the
operational state of this session, having no idea what the uptime or
administrative state is.  In order to be able to at least store an
operational state value, the other attributes of a session need to be
optional.

This therefore makes all the fields optional.