Resolve incoming trap varbinds properly
The varbinds of an incoming trap message were incorrectly resolved (most
likely due to a misunderstanding of PySNMP complexity).

In fact, the original resolver only resolved the name of the incoming
variable, but the subtype of the value also needs to be resolved
according to the MIB definition of the incoming variable.

This new resolver ensures the value is also resolved to the correct
type. An example: The incoming raw value may be an ASN1 OctetString,
but seen in light of the resolved MIB object name, it might actually
be resolved as a InetAddress textual convention type instead (which
would heavily influence how the incoming value is parsed to a Python
value).