Change naive `TrapMessage.variables` definition
`TrapMessage.variables` was naively built as a dictionary to look up
individual trap variables by simple object labels, as we wanted to
easily look up variables in a trap.

However, this implementation turned out to be too naive, because
a trap can contain any number of variables that have the same object
label but different instance numbers (e.g. it can contain all of
`ifOperStatus.1`, `ifOperStatus.2`, `ifOperStatus.3` and so on).  The
naive implementation would overwrite each successive variable in the
variables dict using just the name `ifOperStatus`.

This  introduces a breaking change to the implementation to be a list of
variables, and instead provide convenience methods for common
operations: Verify that an object is present at all in the trap, and
get the full list of all similarly-named objects from the varbinds.