Make trap handling async
Trap handlers are all executed from a synchronous trap callback handler.
This makes it hard for trap handlers to work with async code, like
calling out to make further SNMP queries and wait for the results.

This changes `TrapObserver.handle_trap` to be async, and updates the
`TrapReceiver` class to ensure trap handlers are called asynchronously.