Add option to update single BFD session state
This makes `BFDTask` able to fetch and update the BFD state of a single
session, by adding a `session_index` argument to `LinkState`.

The purpose of this is to enable updating the state of single sessions
on incoming BFD traps: BFD traps do not contain the actual state
value changes, they only contain information about the set of BFD
sessions that changed state to something up-like or something down-like.
The changed state values need to be re-polled, but there is no need
to poll the full table in these cases.