Only log new ifalias values if there is one
Running Zino for the first time against many devices will cause huge
amounts of log messages about interface descriptions being set for the
first time.  There is no reason to log that an interface description
was set to an empty value the first time it was collected, so let's not.

Also, log the descr repr values, just as the code does for changing
values (for consistencty)