Pin `snmpsim` to an older version
A new version of snmpsim was released about 6 hours ago (the first new
version in 5 years), and it seems to have changed to be incompatible
with how our test suite uses it.

In particular, our test suite expects `snmpsimd.py` to be available on
the search PATH, but the executable's name may have changed, and I
don't have the luxury of time to figure out how to adapt our test suite

I assume there may be many more changes, as the original maintainer died
a while back, and someone else has taken over, which is why I think it's
best to pin to an older version for now.