Serialize timedeltas as seconds in legacy protocol
I happened to pick the `ac_down` attribute for testing replacement
of underscores with dashes, which caused me to notice that this
particular timedelta value was not converted to the correct format
expected by the Zino legacy API: Just an integer representing the
number of seconds of the delta.

This adds a regression test for it and fixes the problem.