Translate InetAddress and IpAddress as ip_address
PySNMP doesn't necessarily pretty-print InetAddress and IpAddress
value types as anything useful for us (This are textual conventions
of an OctetString, and PySNMP tends to just raw-dog the octet string
into something like a hexadecimal string representation of the content).

If the value type is `InetAddress` or `IpAddress` we can expect
`ipaddress.ip_address()` to properly parse the raw bytes into an
IPv4Address or IPv6Address object.