Expand log_time_spent decorator
This makes the decorator support async functions.  It also adds a limit
argument which causes the duration log statement to only be emitted if
the decorated function ran for longer than the limit.  An optional
argument formatter function can also be supplied, to format some kind
of useful representation of the incoming function arguments.