Disallow non-int elements in OID
The OID class is designed to represent a tuple of ints.  It was never
designed to contain non-int elements (although it can translate non-int
values to int elements when instances are constructed).

The Zino code has bugs that actually would feed the OID class with
non-int elements, and because the OID class accepted it, strange
workarounds have been written by those who did not understand that
this was not the intended behavior.

This change enforces all elements to be int on instantiation -
presumably at the cost of instantiation becoming slower.