Unroll symbolic PySNMP indices to numeric ones
An `Identifier` was intended to represent an object's MIB name,
symbolic object name and numeric OID index.  However, due to a
misunderstanding of how PySNMP works, this was not always the case:

`getMibSymbol` will actually break down the remaining OID index to
one or more symbolic indices according to the MIB specification, so
the third element of the `getMibSymbol()` call may actually be a tuple
of various objects that represents individual indices.

While this could be helpful in the codebase, it was not the intention
of `_convert_varbind` and `Identifier`.  To keep Zino's interface
as intended here, this change tries to unroll all the symbolic indices
back to a single OID object - likely with some performance penalty.

We might want to look into changing the codebase to work with symbolic
indexes at a later time.