Fix test that was too specific in log checking
I don't generally think it is a good idea to check for very specific
log output.  In this case, the non-int object type returned by the
underlying library may depend on the library.  netsnmpy-cffi does not
generally translate octet strings to `str` unless they have a
textual convention of `DisplayString`.  That has really no bearing on
the actual functionality in this case, so it's quite weird to test
for such specific log messages.