Fix broken interpretation of CISCO bgp peer table
The table index for an address like `10.0.0.42` will be
`.1.4.10.0.0.42`, where `1 is the address type and `4` is the address
length - while the remainder is the address itself.  We don't really
need the `1` and the `4` to detect the address type, as `ip_address`
will parse things correctly if given just a series of bytes.

It turns out that the index unrolling first implemented in
`_convert_varbind()` will cause some elements of the `row_index` OID to
be left out, causing mayhem for client code that relies on these values.