Make fixture trap port configurable
This changes the `localhost_receiver` fixture to use 1163 as its trap
port, and then updates tests that sends external traps to the fixture so
that they will use the port assigned to the fixture.

This will better enable the test suite to change trap ports dynamically,
which might become necessary when running tests for multiple back-ends
in the same test session.