Select unused UDP port for tests
We will be mixing two SNMP libraries in the test suite and cannot
guarantee that expected ports will be free.  We can also never be
sure that port 1162 is always available, so this changes basic
Zino binary tests to utilize the pytest-asyncio fixtured
`unused_udp_port` in order to ensure we found a free port before using
it.