Move dev "requirements" out of pyproject.toml
When specified like this in `pyproject.toml`, these dependencies are
added as package metadata, and will be flagged by some build tools as
optional *runtime* requirements.

These are *not* runtime requirements at all.  A recent PEP has added
non-runtime-requirements as a new type of setting to `pyproject.toml`,
but not all build tools supports it yet, so this moves these
dependencies to `requirements/dev.txt` so developers still have an easy
way to get these tools installed for developing Zino.