Schedule periodic debug logging of SNMP sessions
This schedules periodic debug logging of some stats needed for low level
debugging of SNMP sessions.  I.e. there are multiple levels of SNMP
session objects:

1. Zino's SNMP object (which varies with selected back-end)
2. The underlying library's session object (only applicable for
   netsnmpy)

Zino tries to re-use its own SNMP objects, if possible.  The session
objects underpinning the netsnmpy library must be explicitly closed to
release low-level OS resources (UDP file descriptors, mostly).  This
closing would normally be achieved through garbage collection mechanisms
at the Python level, but something is NOT being garbage collected as
expected, and we are leaking file descriptors because of it.

This periodic log statement should show that the number of re-usable
sessions, the number of reachable (by the gc) session objects and the
number of reachable (by the gc) netsnmpy session are more or less equal.
When they are not, it is a sign of resource leakage.