Replace __del__ with explicit resource management
Using `__del___()` to manage low-level resources seems to be fraught
with potential Heisenbugs, since it may leave resource cleanup up to
the whims of the CPython garbage collector implementation.  Case in
point: It seems to work fine for a running Zino, but wreaks havoc on
the test suite.

This introduces a combination of explicit and implicit resource
management to the `netsnmpy_backend.SNMP` class:  A new pair of
open/close methods are added, as well as making instances into
context managers.  However, since most of Zino does not expect to have
to handle resource allocation/deallocation at this level, implicit
session opening is introduced for all query methods: If the session
is not yet opened, or it has been closed by a previous user but
being reused by something else, the low-level session is opened
when needed.

However, this also means that the Zino job loop needs to explicitly
close SNMP sessions at the end of job runs, and also that polling
functions that are scheduled outside of the job scope must take
care to clean up resources.  This will come in later commits.