Use SNMP context manager outside of job contexts
This updates Task (and some test) code that always or most often runs
outside a normal job context to use explicit SNMP resource management

I.e. these are usually methods that are scheduled to run "later",
when the collection job that scheduled them has ended, and the SNMP
session they refer to has been closed.  Using the SNMP session object
as a context manager here ensures that an SNMP session is opened and
closed before and after these calls.