Imbue ZinoServer with config access
This adds a config attribute and init argument to the ZinoServer class,
so that it may use configuration options to initialize the server
protocol implementations.

This allows for a Configuration object to be injected when creating a
class instance, while omitting it will cause it to use the configuration
from the global state.