Make sparsewalk re-query incomplete chunks
If the last bulk response "chunk" was incomplete (i.e. it did not
contain all the variables we asked for, presumably because there was no
room left in the PDU to fit on the current MTU), the next query
iteration would be based only the response variables present in that
last chunk.  This caused sparsewalk to break completely in such a
situation, causing errors about missing data to be logged.

This change simplifies the sparsewalk routine a bit, by ignoring
a response chunk that is incomplete, and instead issuing the next
bulk query based on the last *complete* response chunk.

This will cause some small amount of redundant data to be transferred,
in some cases, but it is in line with other bulk walk implementations
(like PySNMP's built-in).