Replace pysnmplib with pysnmp
Tests relying on `snmpsim` started failing, likely because of changing 3rd
party dependencies.

Zino2 was stuck on the frozen `pysnmlib` (a fork/snapshot of `pysnmp`
after the author died), because once PySNMP development resumed with new
maintainers, they started changing the APIs.  While the long term plan
is to remove PySNMP support entirely, whatever Zino2 uses needs to be
compatible with `snmpsim` dependencies if we are to continue using
`snmpsim` for the test suite.

It seems we now can get the test suite to run by upgrading from
`pysnmplib` to `pysnmp` 6.2.