Use asyncio PySNMP interface for snmpsim fixture
Updating from pysnmplib to pysnmp 6.2 changes the `nextCmd`
implementation, which now uses asyncio under the hood for the
synchronous interface.  This was incompatible with a test process that
is already running under an asyncio event loop.  The helper functions
needed to explicitly use the async interface in order for the fixture to
run successfully.