Add option to control events for new interfaces
This adds an option to control whether to generate portstate events for
newly discovered interfaces that are operationally down.

The default value reverses the original behavior of Zino, as it
becomes quite tedious (to say the least) for new users when Zino
produces a gajillion events on the first run.