Honor make_events_for_new_interfaces option
The original Zino 1 tweak is to force an interface state to a sentinel
value of UNKNOWN for new interfaces, which later triggers it to create
an event for it when the collected state is operDown but adminUp.

This changes the logic to only apply this tweak if the new
`make_events_for_new_interfaces` option is set to true.

The config object cannot be imported into this module at import time,
as that would cause an empty config object to be imported.  Instead,
the state module is imported, and the config value is copied at runtime
into each instance of the linkstatetask.