Use correct configuration object in main()
Parts of `main()` was referencing the incorrect config object when
loading the configured SNMP back-end. The "official" running config is
in the state module. It is only replaced if `load_config()` returns
an actual config object.  `load_config()` returns an empty result
if the default config file is not present, but raises an exception if
an alternate non-existent config file was specified on the command line.

In order for startup without a config file to work, all config must
be referenced via `state.config`.