Raise exception on illegal event replacement
It was possible to commit identical events, as long as their IDs were
different. This was in fact happening due to a race condition in BFD
event handling: Both the polling task and the trap handler relinquish
control to the event loop between getting a new event and committing,
causing two identical events to be created, but with different IDs.

It should not be possible to use the commit() method to commit two
identically-keyed events with different IDs.  This ensures commit()
raises an exception if this happens.