Create reachability reachable event for new device
Only if config options is set to True

The config object cannot be imported into this module at import time,
as that would cause an empty config object to be imported.  Instead,
the state module is imported, and the config value is copied at runtime
into each instance of the reachabletask.