Use separate SNMP session for extra jobs
Extra reachability queries that run outside of the job context seem
to sometimes hang indefinitely - this is likely an issue because of
race conditions during interleaved usage of low level Net-SNMP sessions.
It may perhaps be better to ask the system for a new session for the
extra job, rather than depend on the old reference of the task object
that initially scheduled the reachability check.