Log unhandled TimeoutErrors in task runs
A TimeoutError would be the exception most expected to occur in a task
run. The remote device stopped responding mid-run (i.e. after the
reachability task determined the device to be reachable), or a single
request was too large to process in a timely manner.  There is no need
to dump a menacing Python traceback whenever this occurs, a single line
will do.

The job will nevertheless run on schedule again soon.