Add module to assist in timeout debug logging
This module adds functionality that will aid in debug logging SNMP
timeout events, from an external viewpoint: This will inspect the
current stack to figure out which Task method timed out, and which SNMP
request timed out (since these things are not attached to the exception
itself). It then debug logs this precise information, rather than
dumping a full traceback (the information is mostly interesting to
developers, which is why this requires enabling debug level logging).