Use microseconds in job name timestamps
Apparently, naming jobs down to the second isn't actually unique
enough: While it is highly unlikely that two POLLRTR requests will
come in from the API at the same second, it is likely that two link-
state traps for the same port will arrive within the same second. The
latter appears to be the case in production, which results in
`apscheduler.jobstores.base.ConflictingIdError` being raised.

This adds microseconds to the timestamp used in a job name, which should
be unique enough for most purposes ;-)