Fix broken `was_flapping` logic
`was_flapping` did not correctly replicate Zino 1 logic.  It was
really supposed to answer whether an interface had been marked as
having crossed the flapping threshold, but instead it only verified
that there were any flap stats registered for an interface.  Flap stats
are *always* kept once a single state change is detected, and are
removed entirely when an interface is "unflapped".

This caused a bug in where the flap ager would incorrectly believe
any recently opened portstate event had transitioned from flapping to
non-flapping, and log this to the event.